fixes after mentors review
